Splitting kmeans.py
